Create new abstract class Crypto that now implement Currency, update 3 classes that extend it
