+++ src/Task1/Bitcoin.php
@@ -3,30 +3,8 @@
 namespace Cryptocurrency\Task1;
 
 
-class Bitcoin implements Currency
+class Bitcoin extends Crypto
 {
-    private $name = 'Bitcoin';
-    private $dailyPrice;
-    private $logoUrl = 'https://s2.coinmarketcap.com/static/img/coins/32x32/1.png';
-
-    public function __construct($dailyPrice)
-    {
-        $this->dailyPrice = $dailyPrice;
-    }
-
-
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    public function getDailyPrice(): float
-    {
-        return $this->dailyPrice;
-    }
-
-    public function getLogoUrl(): string
-    {
-        return $this->logoUrl;
-    }
+    protected $name = 'Bitcoin';
+    protected $logoUrl = 'https://s2.coinmarketcap.com/static/img/coins/32x32/1.png';
 }
